parsers/psense/pdf/pyproject.toml
[tool.poetry]
name = "psense-integ"
version = "0.1.0"
description = "Integration Base for psense.ai"
authors = ["Srikari Rallabandi <sreeku@psense.ai>"]
license = "PSense.ai License"
readme = "README.md"
packages = [{ include = "." }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"  # Set to the Python version you're using
pdfplumber = "^0.11.6"
pytesseract = "^0.3.13"
camelot-py = {extras = ["cv"], version = "^1.0.0"}
pymupdf = "^1.25.5"
scikit-learn = "^1.6.1"
opencv-python = "^4.11.0.86"
opencv-contrib-python = "^4.11.0.86"
pyyaml = "^6.0.2"
transformers = "^4.51.3"
torch = "^2.7.0"
paddleocr = "^2.10.0"
paddlepaddle = "^3.0.0"
symspellpy = "^6.9.0"
fuzzywuzzy = "^0.18.0"
pytest = "^8.3.5"
requests = "^2.32.3"
